While the Post Processing instrumentation is cheap, it is not free. From looking at a lot of shader dumps, I noticed we tend to go

```c
inst_post_process(1, 2, 3);
if (something) 
    inst_post_process(1, 2, 3);
else
    inst_post_process(1, 2, 3);
inst_post_process(1, 2, 3);
```

a lot, so this changes tracks a simple version of Post-Domination in order to detect as much as we can

Using a Godot sample I had running through fossilize with cache off, single threaded, we get quite a good speed up

```
// before
45 graphics pipelines took 23.143 s (accumulated time)
176 compute pipelines took 26.847 s (accumulated time)

// after
45 graphics pipelines took 17.693 s (accumulated time)
176 compute pipelines took 14.548 s (accumulated time)
```

also from the dump we got

```
Function: inst_post_process_descriptor_index
  Max: 215 -> 150
  Mean: 25.88  -> 18.32
  Median: 8.00 -> 7.00
  75th percentile: 15.00 -> 12.00
  95th percentile: 132.10 -> 73.55
```

were you can see for larger shaders we can really reduce quite a bit